+++ src/main.ts
@@ -2,46 +2,12 @@ import _Vue, { VNode } from 'vue';
 import Router from 'vue-router';
 import { DirectiveBinding } from 'vue/types/options';
 
-import { ContentFor, YieldTo } from './components/capture';
-import CmsContent from './components/CmsContent';
+import { Captable } from './api';
 import CmsCarousel from './components/CmsCarousel.vue';
+import CmsContent from './components/CmsContent';
 import CmsZone from './components/CmsZone.vue';
+import { ContentFor, YieldTo } from './components/capture';
 import { addDirectives } from './directives';
-
-export interface Captable {
-}
-
-export interface ZoneCaptable {
-  zoneId?: Captable;
-}
-
-export interface CMSZoneResponse {
-  content: Content[];
-  captable: Captable;
-  zone_type: string;
-  zone_header: string | null;
-  zone_footer: string | null;
-}
-
-export interface HTMLContent {
-  html: string;
-}
-
-export interface Content extends HTMLContent {
-  tracker: string;
-  track_on?: string;
-  tracked?: boolean;
-  extra?: {
-    track_on?: string;
-    external_trackers?: string[];
-  }
-}
-
-export interface ContentTracker {
-  trackOn: string;
-  content: any; // eslint-disable-line @propelinc/no-explicit-any
-  zoneId: string;
-}
 
 interface DestroyHTMLElement extends HTMLElement {
   $destroy: () => void;
@@ -56,12 +22,13 @@ export interface CmsOptions {
   getCaptable?: (() => Captable);
   getSiteVars?: (() => object);
 
-  /* eslint-disable-next-line @propelinc/no-explicit-any */
+  /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
   beforeFetchZone?: null | (() => Promise<any>);
   checkConnection?: (() => boolean);
   onCarouselSwipe?: ((zoneId: string, index: number) => void);
 
-  trackClickHandler?: (eventName: string, eventProps: {[key: string]: any}) => void; // eslint-disable-line @propelinc/no-explicit-any
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  trackClickHandler?: (eventName: string, eventProps: {[key: string]: any}) => void;
 }
 
 export interface PluginOptions extends CmsOptions {
@@ -71,7 +38,9 @@ export interface PluginOptions extends CmsOptions {
   getSiteVars: (() => object);
   globalCssCacheMs: number;
   setCaptable: ((captable: Captable) => void);
-  trackClickHandler: (eventName: string, eventProps: {[key: string]: any}) => void; // eslint-disable-line @propelinc/no-explicit-any
+
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  trackClickHandler: (eventName: string, eventProps: {[key: string]: any}) => void;
 }
 
 export const pluginOptions: PluginOptions = {
@@ -87,7 +56,7 @@ export const pluginOptions: PluginOptions = {
     try {
       const captableStr = localStorage.getItem('captable');
       return JSON.parse(captableStr || '');
-    } catch(e) {
+    } catch (e) {
       return {};
     }
   },
@@ -95,7 +64,7 @@ export const pluginOptions: PluginOptions = {
     return {};
   },
   trackClickHandler(): void {
-    return;
+
   },
 };
 export var finalPluginOptions: PluginOptions;
@@ -115,7 +84,7 @@ const plugin = function install(Vue: typeof _Vue, options?: CmsOptions) {
    * <div v-cms-track-zone="5">Click me</div>
    */
   Vue.directive('cms-track-zone', {
-    bind (el: DestroyHTMLElement, binding: DirectiveBinding, vnode: VNode) {
+    bind(el: DestroyHTMLElement, binding: DirectiveBinding, vnode: VNode) {
       const handler = () => {
         if (vnode.context && vnode.context.$root) {
           vnode.context.$root.$emit(`cms.track.${binding.value}`);
