+++ tests/unit/CmsCarousel.spec.ts
@@ -1,5 +1,5 @@
-import { mount, createLocalVue } from '@vue/test-utils';
 import CmsCarousel from '@/components/CmsCarousel.vue';
+import { mount, createLocalVue } from '@vue/test-utils';
 
 const localVue = createLocalVue();
 
@@ -12,7 +12,7 @@ describe('CmsCarousel.vue', (): void => {
       },
     });
 
-    const component = wrapper.vm as any; // eslint-disable-line @propelinc/no-explicit-any
+    const component = wrapper.vm as any; // eslint-disable-line @typescript-eslint/no-explicit-any
     expect(component.index).toBe(0);
     expect(wrapper.find('.slick-slider').exists()).toBe(true);
     expect(wrapper.find('.slick-current').text()).toMatch('Slide 1');
