+++ tests/unit/CmsContent.spec.ts
@@ -1,8 +1,7 @@
-import { mount, createLocalVue } from '@vue/test-utils';
-import { compileToFunctions } from 'vue-template-compiler';
-import Vue from 'vue';
-
 import CmsContent from '@/components/CmsContent';
+import { mount, createLocalVue } from '@vue/test-utils';
+import Vue from 'vue';
+import { compileToFunctions } from 'vue-template-compiler';
 
 Vue.compile = compileToFunctions;
 const localVue = createLocalVue();
